REFACTOR `PlacesController` into Controllers namespace
